server testing - extract app factory (refs #14)
