+++ src/command/serve.js
@@ -1,10 +1,4 @@
-const express = require('express');
-
-const morgan = require('morgan');
-
-const Renderer = require('../Renderer');
-const SourceDir = require('../SourceDir');
-const Layout = require('../Layout');
+const expressApp = require('../server/expressApp');
 
 /**
  * Serve MD files from rootDir
@@ -14,46 +8,7 @@ const Layout = require('../Layout');
  * @param {String} options.layoutPath path to layout directory
  */
 function serve(options) {
-    const sourceDir = new SourceDir(options.rootDir);
-    const layout = new Layout(options.layoutPath);
-    var renderer = new Renderer(sourceDir, layout, {
-        mode: 'serve',
-    });
-
-    const app = express();
-
-    /*
-     * access logs using apache format
-     * @see https://github.com/expressjs/morgan#predefined-formats
-     */
-    app.use(morgan('tiny'));
-
-    if (layout.hasAssets()) {
-        app.use('/assets', express.static(layout.path + '/assets'));
-    }
-
-    app.get(/^\/(.*)/, function (req, res) {
-        let relativePath = req.params[0];
-        let sourceFile = sourceDir.locateFile(relativePath);
-        if (sourceFile == null) {
-            res.status(404).send('Not found');
-            return;
-        }
-        if (sourceFile.type == 'directory') {
-            let indexFile = sourceDir.locateIndex(sourceFile);
-            if (indexFile == null) {
-                res.status(404).send('Not found');
-                return;
-            }
-            res.send(renderer.render(indexFile));
-            return;
-        }
-        if (['md', 'html'].includes(sourceFile.type)) {
-            res.send(renderer.render(sourceFile));
-        } else {
-            res.sendFile(sourceFile.absolutePath);
-        }
-    });
+    const app = expressApp(options);
 
     app.listen(3000, function () {
         console.log('Application started on http://localhost:3000');
